Some math fixes, and class name ambiguity resolving. solver.* uses ksudoku::SolverState and sksolver.* uses ::SolverState.
svn path=/trunk/KDE/kdegames/ksudoku/; revision=792900

+++ src/logic/skbase.cpp
@@ -35,7 +35,7 @@ void SKBase::setorder(int k, int threedimensionalf, int sized)
 {
 	type = threedimensionalf;
 	order = k; 
-	base = (int) sqrt(order);
+	base = (int) sqrt((double)order);
 	
 	if(sized!=-1) size=sized;
 	else size = (type==1) ?  base*base*base :  (order*order) ;
